HasPreviousCommit causes recursive load of commits unnecessarily
This PR improves HasPreviousCommit to prevent the automatic and recursive loading
of previous commits using git merge-base --is-ancestor and git rev-list

Related #13684
